Correguir refresco de graficos
