+++ src/app/features/products/presentation/components/product-billing-chart/product-billing-chart.component.html
@@ -1,6 +1,19 @@
 <div class="chart-container">
   <div class="chart-header">
     <h3>Facturación Total por Producto</h3>
+    
+    <div class="view-toggle-buttons">
+      <button 
+        [class.active]="billingViewMode === viewModes.ByProduct" 
+        (click)="onBillingViewModeChange(viewModes.ByProduct)">
+        Producto
+      </button>
+      <button 
+        [class.active]="billingViewMode === viewModes.ByProductType" 
+        (click)="onBillingViewModeChange(viewModes.ByProductType)">
+        Tipo de producto
+      </button>
+    </div>
   </div>
   
   <div class="chart-content">
