+++ src/app/features/products/presentation/components/product-sales-chart/product-sales-chart.component.scss
@@ -1,21 +1,45 @@
 .chart-container {
+  background-color: #fff;
+  border-radius: 8px;
+  box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
   padding: 1rem;
   height: 100%;
-  width: 100%;
   display: flex;
   flex-direction: column;
+  width: 100%;
+  box-sizing: border-box;
+  position: relative;
+  margin: 0 auto 0 0;
+}
+
+.chart-header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 1rem;
+  flex-shrink: 0;
+  
+  h3 {
+    margin: 0;
+    font-size: 1.1rem;
+    font-weight: 500;
+    color: #333;
+  }
+}
+
+.chart-content {
+  position: relative;
+  flex: 1;
+  min-height: 400px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
 }
 
 .echarts-container {
-  flex: 1;
-  min-height: 300px;
   width: 100%;
-
-  .chart {
-    height: 100%;
-    min-height: 300px;
-    width: 100%;
-  }
+  height: 100%;
+  min-height: 400px;
 }
 
 .chart-controls {
@@ -69,48 +93,34 @@
   font-size: 1.25rem;
 }
 
-.loading-indicator {
+.loading-state, .error-state, .no-data-state {
   display: flex;
   flex-direction: column;
   align-items: center;
   justify-content: center;
   height: 100%;
-  min-height: 200px;
-  
-  .spinner {
-    width: 40px;
-    height: 40px;
-    border: 4px solid rgba(0, 0, 0, 0.1);
-    border-radius: 50%;
-    border-top-color: #007bff;
-    animation: spin 1s ease-in-out infinite;
-    margin-bottom: 1rem;
-  }
-  
-  @keyframes spin {
-    to { transform: rotate(360deg); }
-  }
+  width: 100%;
+  text-align: center;
+  color: #666;
 }
 
-.error-message {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  height: 100%;
-  min-height: 200px;
-  color: #dc3545;
-  
-  .error-icon {
+.spinner {
+  width: 40px;
+  height: 40px;
+  border: 4px solid rgba(0, 0, 0, 0.1);
+  border-radius: 50%;
+  border-top-color: #333;
+  animation: spin 1s ease-in-out infinite;
+  margin-bottom: 1rem;
+}
+
+@keyframes spin {
+  to { transform: rotate(360deg); }
+}
+
+.error-state, .no-data-state {
+  span {
     font-size: 2rem;
-    font-weight: bold;
-    margin-bottom: 1rem;
-    width: 40px;
-    height: 40px;
-    line-height: 40px;
-    text-align: center;
-    border-radius: 50%;
-    background-color: #dc3545;
-    color: white;
+    margin-bottom: 0.5rem;
   }
 }
